changed transaction amount to double

+++ src/main/java/com/transactions/api/controller/TransactionService.java
@@ -36,7 +36,7 @@ public class TransactionService {
     @POST
     @Path("/createTransaction/{accountId}/{transactionAmount}/{transactionUUID}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response createTransaction(@PathParam("accountId") int accountId, @PathParam("transactionAmount") int transactionAmount, @PathParam("transactionUUID") String transactionUUID) {
+    public Response createTransaction(@PathParam("accountId") int accountId, @PathParam("transactionAmount") double transactionAmount, @PathParam("transactionUUID") String transactionUUID) {
         LOGGER.info("createTransaction is triggered");
         String errorMessageStr = "Internal database error in creating a new transaction for accountId" + accountId;
         return createOrDeleteTransaction(accountId, DBMANAGER.createTransaction(accountId, transactionAmount, transactionUUID), errorMessageStr);
